Added support for the parallel_tests gem.
This update ships a workaround for parallel test execution which was
slowed down or broken due to race conditions on the compiled Apache Avro
schema file repository. Before this update all test workers (threads)
shared the very same file repository which causes race conditions when
one worker performs a refresh and another is trying to lookup schema
definitions. This was fixed with a file repository separation per test
worker.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>